Rename files so that it's the same as UAE
