+++ Makefile.common
@@ -3,7 +3,7 @@ INCFLAGS := -I$(CORE_DIR) \
             -I$(CORE_DIR)/libretro
 
 SOURCES_C += $(CORE_DIR)/libretro/libretro.c
-SOURCES_C += $(CORE_DIR)/libretro/cap32-mapper.c
+SOURCES_C += $(CORE_DIR)/libretro/libretro-mapper.c
 SOURCES_C += $(CORE_DIR)/libretro/diskutils.c
 SOURCES_C += $(CORE_DIR)/libretro/fontmsx.c
 SOURCES_C += $(CORE_DIR)/libretro/graph.c
