+++ libretro/diskutils.c
@@ -1,4 +1,4 @@
-#include "libretro-cap32.h"
+#include "libretro-glue.h"
 #include "graph.h"
 
 #include <unistd.h>
