+++ libretro/libretro-glue.h
