+++ libretro/vkbd.c
@@ -1,4 +1,4 @@
-#include "libretro-cap32.h"
+#include "libretro-glue.h"
 
 #include "vkbd_def.h"
 #include "graph.h"
